fix: use data binding correctly for signin section
