+++ app/src/main/java/com/example/logintask1/signinPackage/SigninFragment.kt
@@ -1,13 +1,15 @@
-package com.example.logintask1
+package com.example.logintask1.signinPackage
 
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
 import android.view.View
+import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
+import com.example.logintask1.R
 import com.example.logintask1.databinding.FragmentSigninBinding
 import com.google.android.material.textfield.TextInputEditText
 import com.google.android.material.textfield.TextInputLayout
@@ -23,19 +25,23 @@ class SigninFragment : Fragment(R.layout.fragment_signin) {
 
         binding = FragmentSigninBinding.bind(view)
         binding.viewModel = model
+        binding.lifecycleOwner = this
 
-        observe(binding.emailContainer, model.emailError)
-        observe(binding.passwordContainer, model.passwordError)
-        addChangeListener(binding.emailEditText, model::updateEmailError)
-        addChangeListener(binding.passwordEditText, model::updatePasswordError)
+
+
+//        observe(binding.emailContainer, model.emailError)
+//        observe(binding.passwordContainer, model.passwordError)
+//        addChangeListener(binding.emailEditText, model::updateEmailError)
+//        addChangeListener(binding.passwordEditText, model::updatePasswordError)
 
     }
 
-    private fun observe(container: TextInputLayout, memberToObserve: MutableLiveData<String?>){
-        memberToObserve.observe(viewLifecycleOwner,  Observer<String?> { text ->
-            container.helperText = text
-        })
-    }
+    // TODO: remove this
+//    private fun observe(container: TextInputLayout, memberToObserve: MutableLiveData<String?>){
+//        memberToObserve.observe(viewLifecycleOwner,  Observer<String?> { text ->
+//            container.helperText = text
+//        })
+//    }
     private fun addChangeListener(container: TextInputEditText, modelViewFunction: KFunction1<String, Unit>){
         container.addTextChangedListener(object: TextWatcher {
             override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
